refactor code (json parser construct into json serializer and deserializer)
