+++ src/main/java/ru/clevertec/Main.java
@@ -4,7 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import ru.clevertec.config.JacksonConfig;
 import ru.clevertec.domain.Customer;
 import ru.clevertec.util.DataGenerator;
-import ru.clevertec.util.JsonParser;
+import ru.clevertec.util.JsonDeserializer;
+import ru.clevertec.util.JsonSerializer;
 import ru.clevertec.util.io.FileReader;
 import ru.clevertec.util.io.FileWriter;
 
@@ -15,16 +16,17 @@ public class Main {
 
         ObjectMapper objectMapper = JacksonConfig.objectMapper();
         Customer customer = DataGenerator.generateRandomCustomer();
-        JsonParser parser = new JsonParser();
+        JsonSerializer serializer = new JsonSerializer();
+        JsonDeserializer deserializer = new JsonDeserializer();
 
-        StringBuilder sb = parser.toJson(customer);
+        StringBuilder sb = serializer.toJson(customer);
         sb.append("}");
         FileWriter.writeToFile(FILEPATH_TO_CUSTOM_OUTPUT, sb.toString());
         FileWriter.writeToFile(FILEPATH_TO_JACKSON_OUTPUT, objectMapper.writeValueAsString(customer));
 
         String json = FileReader.readFromFile(FILEPATH_TO_CUSTOM_OUTPUT);
 
-        Customer resultCustomer = parser.fromJson(json, Customer.class);
+        Customer resultCustomer = deserializer.fromJson(json, Customer.class);
         Customer jacksonCustomer = objectMapper.readValue(json, Customer.class);
 
         System.out.println("myCustomer after toJson and from json transformation: " + resultCustomer);
