+++ src/main/java/ru/clevertec/util/jackson/deserializer/LocalDateTimeDeserializer.java
@@ -1,24 +1,25 @@
-package ru.clevertec.util.deserializer;
+package ru.clevertec.util.jackson.deserializer;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 
 import java.io.IOException;
-import java.time.OffsetDateTime;
+import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class OffsetDateTimeDeserializer extends JsonDeserializer<OffsetDateTime> {
+public class LocalDateTimeDeserializer extends JsonDeserializer<LocalDate> {
     private static final DateTimeFormatter DATE_TIME_FORMATTER
-            = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.nnnnnnnXXXXX");
+            = DateTimeFormatter.ofPattern("yyyy-MM-dd");
 
     @Override
-    public OffsetDateTime deserialize(JsonParser jsonParser, DeserializationContext deserializationContext)
+    public LocalDate deserialize(JsonParser jsonParser, DeserializationContext deserializationContext)
             throws IOException {
         String dateAsString = jsonParser.getText();
         if (dateAsString == null) {
             throw new IOException("OffsetDateTime argument is null.");
         }
-        return OffsetDateTime.parse(dateAsString, DATE_TIME_FORMATTER);
+        return LocalDate.parse(dateAsString, DATE_TIME_FORMATTER);
     }
 }
+
