+++ src/main/java/ru/clevertec/util/jackson/deserializer/OffsetDateTimeDeserializer.java
@@ -1,25 +1,24 @@
-package ru.clevertec.util.deserializer;
+package ru.clevertec.util.jackson.deserializer;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 
 import java.io.IOException;
-import java.time.LocalDate;
+import java.time.OffsetDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class LocalDateTimeDeserializer extends JsonDeserializer<LocalDate> {
+public class OffsetDateTimeDeserializer extends JsonDeserializer<OffsetDateTime> {
     private static final DateTimeFormatter DATE_TIME_FORMATTER
-            = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+            = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.nnnnnnnXXXXX");
 
     @Override
-    public LocalDate deserialize(JsonParser jsonParser, DeserializationContext deserializationContext)
+    public OffsetDateTime deserialize(JsonParser jsonParser, DeserializationContext deserializationContext)
             throws IOException {
         String dateAsString = jsonParser.getText();
         if (dateAsString == null) {
             throw new IOException("OffsetDateTime argument is null.");
         }
-        return LocalDate.parse(dateAsString, DATE_TIME_FORMATTER);
+        return OffsetDateTime.parse(dateAsString, DATE_TIME_FORMATTER);
     }
 }
-
