+++ src/main/java/ru/clevertec/util/jackson/serializer/OffsetDateTimeSerializer.java
@@ -1,4 +1,4 @@
-package ru.clevertec.util.serializer;
+package ru.clevertec.util.jackson.serializer;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
